update the scripter casting

+++ src/main/java/com/kytheralabs/SeleniumScripter.java
@@ -101,7 +101,7 @@ public class SeleniumScripter {
     private void iterateTable(Map<String, Object> script) throws Exception {
         int offset = 0;
         if(script.containsKey("rowoffset")){
-            offset = ((Integer)script.get("rowoffset")).intValue();
+            offset = ((Double) script.get("rowoffset")).intValue();
         }
         while (true) {
             List<WebElement>  allRows = selectElements(script.get("selector").toString(), script.get("name").toString());
